+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFE1MXAKT8YB63RVW4BT69N7"
+        "01KFF8M4VEMAG0YNPSESQ84A44"
     )
     r"""The unique identifier of the data source"""