+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFE1MXAH29P3YAC300SDQ8PM"
+        "01KFF8M4VA30XFRNWTTWBXZN8T"
     )
     r"""The unique identifier of the data source"""