+++ packages/orq-rc/RELEASES.md
@@ -4899,3 +4899,13 @@ Based on:
 - [python v4.2.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v4.2.0-rc.52] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.52 - packages/orq-rc
+
+## 2026-01-21 03:14:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.691.0 (2.797.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.2.0-rc.53] packages/orq-rc
+### Releases
+- [PyPI v4.2.0-rc.53] https://pypi.org/project/orq-ai-sdk/4.2.0-rc.53 - packages/orq-rc